+++ src/execution/executor_index_scan.h
@@ -139,7 +139,7 @@ class IndexScanExecutor : public AbstractExecutor {
                 } else if (min_value.type == TYPE_INT) {
                     min_value.init_raw(sizeof(int));
                 } else if (min_value.type == TYPE_FLOAT) {
-                    min_value.init_raw(sizeof(float));
+                    min_value.init_raw(sizeof(double));
                 }
             }
             if (max_value.raw == nullptr) {
@@ -148,7 +148,7 @@ class IndexScanExecutor : public AbstractExecutor {
                 } else if (max_value.type == TYPE_INT) {
                     max_value.init_raw(sizeof(int));
                 } else if (max_value.type == TYPE_FLOAT) {
-                    max_value.init_raw(sizeof(float));
+                    max_value.init_raw(sizeof(double));
                 }
             }
             memcpy(upper_record.data + offset, max_value.raw->data, col.len);
